refactor(edge): isolate external auth flow and header mutation logic
